Add User superclass for Employee and Customer to integrate with Spring Security.
