+++ backend/src/main/java/com/cathy/shopping/model/User.java
@@ -1,0 +1,135 @@
+package com.cathy.shopping.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+import java.time.LocalDate;
+import java.util.Arrays;
+
+@MappedSuperclass
+public abstract class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    @NotBlank(message = "Username is mandatory")
+    private String username;
+
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    @NotBlank(message = "Password is mandatory")
+    private String password;
+
+    @Column(name = "display_name")
+    @NotBlank(message = "Display name is mandatory")
+    private String displayName;
+
+    @Column(unique = true)
+    @NotBlank(message = "Email is mandatory")
+    @Email
+    private String email;
+
+    private String address;
+
+    @Column(length = 2097152)
+    @Lob
+    @Basic(fetch = FetchType.LAZY)
+    private byte[] photo;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Role role;
+
+    @Column(nullable = false)
+    private String source;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public @NotBlank(message = "Username is mandatory") String getUsername() {
+        return username;
+    }
+
+    public void setUsername(@NotBlank(message = "Username is mandatory") String username) {
+        this.username = username;
+    }
+
+    public @NotBlank(message = "Password is mandatory") String getPassword() {
+        return password;
+    }
+
+    public void setPassword(@NotBlank(message = "Password is mandatory") String password) {
+        this.password = password;
+    }
+
+    public @NotBlank(message = "Display name is mandatory") String getDisplayName() {
+        return displayName;
+    }
+
+    public void setDisplayName(@NotBlank(message = "Display name is mandatory") String displayName) {
+        this.displayName = displayName;
+    }
+
+    public @NotBlank(message = "Email is mandatory") @Email String getEmail() {
+        return email;
+    }
+
+    public void setEmail(@NotBlank(message = "Email is mandatory") @Email String email) {
+        this.email = email;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public byte[] getPhoto() {
+        return photo;
+    }
+
+    public void setPhoto(byte[] photo) {
+        this.photo = photo;
+    }
+
+    public Role getRole() {
+        return role;
+    }
+
+    public void setRole(Role role) {
+        this.role = role;
+    }
+
+    public String getSource() {
+        return source;
+    }
+
+    public void setSource(String source) {
+        this.source = source;
+    }
+
+    @Override
+    public String toString() {
+        return "User{" +
+                "id=" + id +
+                ", username='" + username + '\'' +
+                ", password='" + password + '\'' +
+                ", displayName='" + displayName + '\'' +
+                ", email='" + email + '\'' +
+                ", address='" + address + '\'' +
+                ", photo=" + Arrays.toString(photo) +
+                ", role=" + role +
+                ", source='" + source + '\'' +
+                '}';
+    }
+}
